2080_LLM/DataAugmentation_ReDSM5/requirements.txt
black
hydra-core==1.3.2
jsonlines
matplotlib
nlpaug
numpy
optuna
pandas
pytest
python-dateutil
PyYAML
scikit-learn
textattack
transformers
ruff
mypy
accelerate
tqdm
mlflow>=2.9.0
dvc>=3.0.0
dvc-s3
psycopg2-binary
tensorboard
optuna-dashboard
wandb
